added entity delete constraints

+++ app/blueprints/departments.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash
-from models import db, Department
+from models import db, Department, Position, Contractor
 
 departments_bp = Blueprint('departments', __name__)
 
@@ -35,6 +35,14 @@ def update(department_id):
 @departments_bp.route('/departments/delete/<int:department_id>', methods=['POST'])
 def delete(department_id):
     dept = Department.query.get_or_404(department_id)
+    pos_count = Position.query.filter_by(department_id=department_id).count()
+    if pos_count > 0:
+        flash('Cannot delete department: it has related positions.')
+        return redirect(url_for('departments.list'))
+    contractor_count = Contractor.query.filter_by(department_id=department_id).count()
+    if contractor_count > 0:
+        flash('Cannot delete department: it has assigned contractors.')
+        return redirect(url_for('departments.list'))
     db.session.delete(dept)
     db.session.commit()
     flash('Department deleted!')
